Resurrect ConnectedGothamService as InetGothamService
Experimenting with the exposed service in a "tower middleware" type of
use case (not just an "alternate transport" use case) indicated that
GothamService::connect really should have been public to allow
third-party integrators the same power as the provided start functions.
At the same time, the actual lifetime of these service objects within
Hyper has become more clear to me, and the previous division between
GothamService and ConnectedGothamService makes sense.

On the other hand it's still true that not all transports will have a
SoketAddr to provide, so in this version GothamService and
InetGothamService both implement HttpService separately.